move TextScores to types_.py and convert string_to_model_inputs to a factory method on ModelInputs
