+++ src/neural_tokenizers/mistral_segmenter.py
@@ -11,6 +11,7 @@ from neural_tokenizers.types_ import (
     ModelInputs,
     ScoredTokenIds,
     Segment,
+    TextScores,
 )
 from neural_tokenizers.utils import (
     LN_OF_TWO,
@@ -18,72 +19,6 @@ from neural_tokenizers.utils import (
     lengths_to_float_attention_mask_right_padding,
     pool_segments_by_lengths,
 )
-
-
-@dataclass
-class TextScores:
-    """A segmentation of a string into segments of characters/tokens."""
-
-    text: str
-    chars: list[str]
-    char_lens: LongTensor
-    byte_ids: LongTensor
-    attention_mask: BoolTensor | None = None
-    # character score (composite score)
-    char_scores: FloatTensor | None = None
-    char_surprisals: FloatTensor | None = None
-    char_entropies: FloatTensor | None = None
-    # byte-level information-theoretic scores
-    byte_surprisals: FloatTensor | None = None
-    byte_entropies: FloatTensor | None = None
-
-    @classmethod
-    def from_string(cls, text: str) -> "TextScores":
-        """Create a TextSegmentation from a string.
-
-        Args:
-            text: str
-
-        Returns:
-            TextSegmentation
-
-        """
-        model_inputs = string_to_model_inputs(text)
-        # ∷ (B × L)
-        byte_ids = model_inputs.input_ids
-        # ∷ (B × L)
-        attention_mask = model_inputs.attention_mask
-        chars = list(text)
-        # ∷ (L) → (B × L)
-        char_lens = torch.tensor([len(c.encode("utf-8")) for c in chars]).unsqueeze(0)
-        return cls(
-            text=text,
-            chars=chars,
-            char_lens=char_lens,
-            byte_ids=byte_ids,
-            attention_mask=attention_mask,
-        )
-
-
-def string_to_model_inputs(text: str) -> ModelInputs:
-    """Convert a string into a model input (for transformers-compatible models).
-
-    Args:
-        text: str
-
-    Returns:
-        ModelInputs
-
-    """
-    ids = torch.tensor(list(text.encode("utf-8")), dtype=torch.long).unsqueeze(0)
-    length = len(ids)
-    causal_mask_bool = torch.tril(torch.ones(length, length, dtype=torch.bool))
-    # ∷ (Batch × Heads × Length × Length), where Batch=1, Heads=1
-    causal_mask_bool = causal_mask_bool.unsqueeze(0).unsqueeze(1)
-    causal_mask_float_additive = torch.zeros_like(
-        causal_mask_bool, dtype=torch.float32
-    ).masked_fill(~causal_mask_bool, -float("inf"))
-    return ModelInputs(input_ids=ids, attention_mask=causal_mask_float_additive)
 
 
 def encode_batch(texts: list[str]) -> ModelInputs:
